better types, placement
