+++ frontend/src/metabase/collections/utils.ts
@@ -347,22 +347,3 @@ export const getCollectionPathAsString = (collection: CollectionEssentials) => {
 };
 
 export const collectionPathSeparator = "/";
-
-/**
- * Returns true if the item is the collection itself or a descendant of it.
- * Used to prevent circular references when moving or editing collections.
- */
-export function isItemInCollectionOrItsDescendants(
-  item: Pick<CollectionItem, "id" | "effective_location" | "location">,
-  collectionId: CollectionId | undefined,
-): boolean {
-  if (collectionId === undefined) {
-    return false;
-  }
-
-  const location = item.effective_location ?? item.location;
-  return (
-    item.id === collectionId ||
-    location?.split("/").includes(String(collectionId)) === true
-  );
-}
