+++ frontend/src/metabase/collections/utils.unit.spec.ts
@@ -4,7 +4,6 @@ import {
   getCollectionPathAsString,
   isExamplesCollection,
   isItemCollection,
-  isItemInCollectionOrItsDescendants,
   isReadOnlyCollection,
   isRootCollection,
   isRootPersonalCollection,
@@ -244,72 +243,6 @@ describe("Collections > utils", () => {
       });
       const pathString = getCollectionPathAsString(collection);
       expect(pathString).toBe("Home / User / Files / Documents");
-    });
-  });
-
-  describe("isItemInCollectionOrItsDescendants", () => {
-    it("returns false when collectionId is undefined", () => {
-      const item = createMockCollectionItem({
-        id: 1,
-        effective_location: "/2/3",
-        location: "/2/3",
-      });
-      expect(isItemInCollectionOrItsDescendants(item, undefined)).toBe(false);
-    });
-
-    it("returns true when item.id matches collectionId", () => {
-      const item = createMockCollectionItem({
-        id: 5,
-        effective_location: "/2/3",
-        location: "/2/3",
-      });
-      expect(isItemInCollectionOrItsDescendants(item, 5)).toBe(true);
-    });
-
-    it("returns false when item.id does not match and is not a descendant", () => {
-      const item = createMockCollectionItem({
-        id: 10,
-        effective_location: "/2/3",
-        location: "/2/3",
-      });
-      expect(isItemInCollectionOrItsDescendants(item, 5)).toBe(false);
-    });
-
-    it("returns true when item is a descendant (collectionId in effective_location)", () => {
-      const item = createMockCollectionItem({
-        id: 10,
-        effective_location: "/2/5/8",
-        location: "/2/5/8",
-      });
-      expect(isItemInCollectionOrItsDescendants(item, 5)).toBe(true);
-    });
-
-    it("returns true when item is a nested descendant", () => {
-      const item = createMockCollectionItem({
-        id: 20,
-        effective_location: "/1/2/3/4/5",
-        location: "/1/2/3/4/5",
-      });
-      expect(isItemInCollectionOrItsDescendants(item, 3)).toBe(true);
-    });
-
-    it("uses location as fallback when effective_location is undefined", () => {
-      const item = createMockCollectionItem({
-        id: 10,
-        effective_location: undefined,
-        location: "/2/5/8",
-      });
-      expect(isItemInCollectionOrItsDescendants(item, 5)).toBe(true);
-    });
-
-    it("does not match partial IDs in location path", () => {
-      const item = createMockCollectionItem({
-        id: 100,
-        effective_location: "/12/123/1234",
-      });
-      expect(isItemInCollectionOrItsDescendants(item, 1)).toBe(false);
-      expect(isItemInCollectionOrItsDescendants(item, 12)).toBe(true);
-      expect(isItemInCollectionOrItsDescendants(item, 123)).toBe(true);
     });
   });
 });
