+++ frontend/src/metabase/common/components/Pickers/utils.ts
@@ -240,3 +240,22 @@ export const canCollectionCardBeUsed = (
 
   return true;
 };
+
+/**
+ * Returns true if the item is the collection itself or a descendant of it.
+ * Used to prevent circular references when moving or editing collections.
+ */
+export function isItemInCollectionOrItsDescendants(
+  item: Pick<CollectionPickerItem, "id" | "effective_location" | "location">,
+  collectionId: CollectionId | undefined,
+): boolean {
+  if (collectionId === undefined) {
+    return false;
+  }
+
+  const location = item.effective_location ?? item.location;
+  return (
+    item.id === collectionId ||
+    location?.split("/").includes(String(collectionId)) === true
+  );
+}
